add app ml_models, restructuring
