+++ backend/licenses/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from users.models import HealthFirstUser
 from django.utils import timezone
-from licenses.utils import file_utils
+from ml_models.utils import file_utils
 from django.core.exceptions import ValidationError
 
 # Create your models here.
@@ -153,38 +153,3 @@ class Certificate(models.Model):
             keys=[self.license.user.first_name,self.license.user.last_name,str(self.license.user.dni)] #ojo con dni con punto, se queda con las palabras claves para ownership
             return file_utils.search_in_pdf_text(keys,certificate_text) and file_utils.date_in_range(certificate_text,self.license) #si encontró las palabras claves y una fecha que en el certificado que entra en rango
         return False
-
-
-class LicenseDatasetEntry(models.Model):
-    GROUP_CHOICES = [
-        ('accidente_trabajo', 'Accidente de trabajo'),
-        ('donacion_sangre', 'Donación de sangre'),
-        ('duelo', 'Duelo tipo A y B'),
-        ('enfermedad', 'Enfermedad'),
-        ('asistencia_familiares', 'Asistencia de familiares'),
-        ('estudios', 'Estudios'),
-        ('matrimonial', 'Matrimonial'),
-        ('mudanza', 'Mudanza'),
-        ('gremial', 'Gremial'),
-        ('nacimiento', 'Nacimiento'),
-        ('salud_materna', 'Salud Materna'),
-        ('vacaciones', 'Vacaciones'),
-        ('otro', 'Otro'),
-    ]
-
-    STATUS_CHOICES = [
-        ('approved', 'Aprobado'),
-        ('rejected', 'Rechazado'),
-    ]
-
-    text = models.TextField()
-    type = models.CharField(max_length=30, choices=GROUP_CHOICES)
-    status = models.CharField(max_length=10, choices=STATUS_CHOICES)
-    reason = models.TextField(blank=True, null=True)
-
-    def clean(self):
-        if self.status == 'rejected' and not self.reason:
-            raise ValidationError('Debe especificar un motivo si el estado es "rejected".')
-
-    def __str__(self):
-        return f"{self.type} ({self.status})"
