+++ backend/ml_models/anomalies/isolation_forest_sup_model.pkl
