+++ backend/ml_models/anomalies/isolation_forest_sup_model_v3.pkl
