+++ backend/ml_models/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class MlModelsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'ml_models'
