+++ backend/users/views.py
@@ -14,7 +14,7 @@ from urllib.parse import unquote
 from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.permissions import IsAuthenticated
 from rest_framework_simplejwt.authentication import JWTAuthentication
-from licenses.utils.file_utils import *
+from ml_models.utils.file_utils import *
 from messaging.services.brevo_email import *
 from .health_risk.risk_model import predict_employ_risk, predict_risk
 from django.views.decorators.csrf import csrf_exempt
